refactor(cli): co-locate built-in command definitions
Move each command's CliCommandDefinition next to its implementation and register them via builtins.ts instead of a central declarations file.
